Fixed comments on review
